% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{numeric value for number of trials}

\item{k}{numeric value for number of successes}
}
\value{
combinations of successes
}
\description{
bin_choose
}
\examples{
#number of combinations in which 2 successes occur in 5 trials
bin_choose(n = 5, k = 2)

#number of combinations in which 1:3 successes occur in 5 trials
bin_choose(5, 1:3)
}
