% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of one success}
}
\value{
table of success rates
}
\description{
bin_distribution
}
