% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of one success}
}
\value{
"tailedness" of the probility disribution
}
\description{
bin_kurtosis
}
