% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{numeric value for number of successes}

\item{n}{numeric value for number of trials}

\item{p}{numeric value for probability of one success}
}
\value{
probability of k successes
}
\description{
bin_probability
}
\examples{
#probability of 2 successes in 5 trials with success rate of 0.5
bin_probability(2, 5, 0.5)

#probability of 0:2 successes in 5 trials with success rate of 0.5
bin_probability(0:2, 5, 0.5)

#probability of 55 successes in 100 trials with success rate of 0.45
bin_probability(55, 100, .45)
}
