% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of one success}
}
\value{
asymmetry of probability distribution
}
\description{
bin_skewness
}
