% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of one success}
}
\value{
object with class "binvar"
}
\description{
bin_variable
}
