% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Workout3.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of one success}
}
\value{
variance of number of successes
}
\description{
bin_variance
}
