% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sample_exercise.R
\name{sample_exercise}
\alias{sample_exercise}
\title{sample_exercise}
\usage{
sample_exercise(ex_list)
}
\arguments{
\item{ex_list}{A list of exercises to be sampled from.}
}
\value{
`ex_sample` A vector containing 5 sampled exercises.
}
\description{
A function designed to sample from a list of exercises grouped by type to produce a more rounded program.
}
